added modifications for login and admin
