+++ main.py
@@ -1,5 +1,5 @@
-# import dbOperation
-import adminMenu1
+import adminOperations
+import adminMenu
 
 noExit = True
 
@@ -8,7 +8,7 @@ while noExit:
     try:
         choice = int(input("Select an option: "))
     except ValueError:
-        print('\nWrong selection!! Try again.')
+        print('\nWrong selection!! Try again.', ValueError)
         continue
 
     if choice == 3:
@@ -17,15 +17,18 @@ while noExit:
     else:
         loginId = input("Enter login Id: ").lower()
         password = input("Enter password: ")
-        # user = dbOperation.checkLoginId(loginId)
-        # user = adminOperations.checkLoginId(loginId) # generic function can go in a different py file?
-
-        user = True #to pass the validation
+        user = adminOperations.check_login_id(loginId)
+        print("user", user)
 
         if not user:  # User not found
             print('User not found!')
-        elif choice == 1:  # admin(1)
-            adminMenu1.adminMenu()
-        else:  # customer(2)
-            print("Show customer options")  # remove this after adding code for customer
-        #   Create new file, import and call function here to display the customer options
+        else:
+            userAuth = adminOperations.checkPswd(loginId, password)
+            if(userAuth):
+                if choice == 1:  # admin(1)
+                    adminMenu.adminMenu()
+                else:  # customer(2)
+                    print("Show customer options")  # remove this after adding code for customer
+                #   Create new file, import and call function here to display the customer options
+            else:
+                print('Wrong password!')
